src/backend/requirement.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Werkzeug==3.0.1
python-dotenv==1.0.1
requests==2.32.3
